% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/any_non_na.R
\name{any_non_na}
\alias{any_non_na}
\title{A function to test if a selected group of cols has any NAs}
\usage{
any_non_na(x)
}
\arguments{
\item{mypkg}{Name of package to check}
}
\description{
This function has one arg.
}
\examples{
any_non_na(x)
}
\keyword{NA}
\keyword{NA,}
\keyword{any}

