% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chars_to_datetimes.R
\name{chars_to_datetimes}
\alias{chars_to_datetimes}
\title{A date Function}
\usage{
chars_to_datetimes(df, date_cols, format)
}
\arguments{
\item{df}{dataframe}

\item{date_cols}{lists the date columns}

\item{format}{}
}
\description{
This function allows you to convert char dates to datetimes.
}
\examples{
chars_to_datetimes()
}
\keyword{dates}

