% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compress_ph.R
\name{compress_ph}
\alias{compress_ph}
\title{A function to remove a specific char from a field.}
\usage{
compress_ph(df, x, val2del = "")
}
\arguments{
\item{mypkg}{Name of package to check}
}
\description{
This function takes 3 args...the data frame, the field, the value to delete.
}
\examples{
compress_ph(df, x, val2del)
}
\keyword{compression,}
\keyword{substitution}

