% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/concat_cols.R
\name{concat_cols}
\alias{concat_cols}
\title{A function to to concat data frame columns.}
\usage{
concat_cols(df, cols, newcol)
}
\arguments{
\item{mypkg}{Name of package to check}
}
\description{
This function will concatenate any number of columns into one.
}
\examples{
concat_cols(df, cols, newcol)
}
\keyword{concatenation}

