% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/copy_tactic_ids.R
\name{copy_tactic_ids}
\alias{copy_tactic_ids}
\title{A function to copy tactic ids.}
\usage{
copy_tactic_ids(x)
}
\arguments{
\item{mypkg}{Name of package to check}
}
\description{
This function takes 1 arg...a data frame.
}
\examples{
copy_tactic_ids(x)
}
\keyword{copy,}
\keyword{ids}
\keyword{tactic}

