% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fill_in_tids.R
\name{fill_in_tids}
\alias{fill_in_tids}
\title{A function to fill in tactic ids.}
\usage{
fill_in_tids(x)
}
\arguments{
\item{mypkg}{Name of package to check}
}
\description{
This function takes 1 arg...a data frame
}
\examples{
fill_in_tids(x)
}
\keyword{id}
\keyword{tactic}

