% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_fpaths.R
\name{get_fpaths}
\alias{get_fpaths}
\title{Retrieve full file path for a directory.}
\usage{
get_fpaths(p, pat)
}
\arguments{
\item{p}{which is file path, and pat, to indicate the file type.}
}
\description{
This function takes 2 arguments
}
\examples{
get_fpaths(p, pat)
}
\keyword{file}
\keyword{path,}
\keyword{url}

