% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hrs_to_secs.R
\name{hrs_to_secs}
\alias{hrs_to_secs}
\title{Converts hours to seconds.}
\usage{
hrs_to_secs(h)
}
\arguments{
\item{h}{is the number of hours to be converted to secs.}
}
\description{
This function takes 1 argument.
}
\examples{
hrs_to_secs(h)
}
\keyword{conversion}
\keyword{hours,}
\keyword{seconds,}

