% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is.installed}
\alias{is.installed}
\title{A function to check if a package is installed}
\usage{
is.installed(mypkg)
}
\arguments{
\item{mypkg}{Name of package to check}
}
\description{
This function allows you to check whether an r package is installed
}
\examples{
is.installed()
}
\keyword{installation}

