% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/match_write_tid.R
\name{match_write_tid}
\alias{match_write_tid}
\title{The main function for this project. Looks up values between spreadsheets.}
\usage{
match_write_tid(x, y, z = "", xcol = "", ycol = "")
}
\arguments{
\item{mypkg}{Name of package to check}
}
\description{
This function allows you to match columns and bring in a field.
}
\examples{
match_write_tid()
}
\keyword{match}

