% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r_filename.R
\name{r_filename}
\alias{r_filename}
\title{This will tell you give you the basepart of the filename.}
\usage{
r_filename(x)
}
\arguments{
\item{p}{which is file path, and pat, to indicate the file type.}
}
\description{
This function takes 1 argument
}
\examples{
r_filename(x)
}
\keyword{file}
\keyword{path,}
\keyword{url}

