% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stringManip.r
\name{right}
\alias{right}
\title{A function to extract a string by starting with the right side of the string.}
\usage{
right(string, char)
}
\arguments{
\item{mypkg}{Name of package to check}
}
\description{
This function has two args.
}
\examples{
right(x, y)
}
\keyword{NA}
\keyword{NA,}
\keyword{any}

