% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/to_date.R
\name{to_date}
\alias{to_date}
\title{Takes a dataframe and a vector of columns and converts to dates}
\usage{
to_date(df, cols)
}
\arguments{
\item{df}{which is a data frame, and cols which is columns.}
}
\description{
This function takes 2 arguments
}
\examples{
r_filename(x)
}
\keyword{file}
\keyword{path,}
\keyword{url}

